Add a builtin to construct a `std::initializer_list` from an array.
