+++ toolchain/check/eval.cpp
@@ -1784,7 +1784,8 @@ static auto MakeConstantForBuiltinCall(EvalContext& eval_context,
     case SemIR::BuiltinFunctionKind::IntRightShiftAssign:
     case SemIR::BuiltinFunctionKind::PointerMakeNull:
     case SemIR::BuiltinFunctionKind::PointerIsNull:
-    case SemIR::BuiltinFunctionKind::PointerUnsafeConvert: {
+    case SemIR::BuiltinFunctionKind::PointerUnsafeConvert:
+    case SemIR::BuiltinFunctionKind::CppStdInitializerListMake: {
       // These are runtime-only builtins.
       // TODO: Consider tracking this on the `BuiltinFunctionKind`.
       return SemIR::ConstantId::NotConstant;
