+++ toolchain/sem_ir/BUILD
@@ -85,6 +85,7 @@ cc_library(
         "builtin_function_kind.cpp",
         "class.cpp",
         "constant.cpp",
+        "cpp_initializer_list.cpp",
         "facet_type_info.cpp",
         "file.cpp",
         "function.cpp",
@@ -106,6 +107,7 @@ cc_library(
         "constant.h",
         "copy_on_write_block.h",
         "cpp_global_var.h",
+        "cpp_initializer_list.h",
         "cpp_overload_set.h",
         "entity_name.h",
         "entity_with_params_base.h",
